requirements.txt
numpy
pandas
sounddevice
scipy
pyaudio
openai-whisper
ffmpeg-python
ollama
langchain
langchain-community
faiss-cpu
sentence-transformers
pyttsx3
selenium
webdriver-manager
pyautogui
requests
python-dotenv
jsonschema
transformers
torch
speechbrain

